git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4417 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/sandbox/learnbip/Experiments.cpp
@@ -113,6 +113,6 @@ Experiments::Experiments(rwlibs::task::CartesianTask::Ptr experiments){
         _experiments[i].mark				= experiment->getPropertyMap().get<int>("Mark",-1);
         _experiments[i].poseEstimateError	= experiment->getPropertyMap().get<double>("PoseEstimateError",-1);
         _experiments[i].inlierFraction		= experiment->getPropertyMap().get<double>("InlierFraction",-1);
-        _experiments[i].temperature			= experiment->getPropertyMap().get<std::vector<double> >("Temperature",std::vector<int>());
+        _experiments[i].temperature			= experiment->getPropertyMap().get<std::vector<double> >("Temperature",std::vector<double>());
     }
 }
